Created Sidebar and new routes

+++ src/actions/auth.js
@@ -1,11 +1,18 @@
 import axios from "../helpers/axios";
-import { LOGIN_FAILURE, LOGIN_REQUEST, LOGIN_SUCCESS, LOGOUT_REQUEST } from "./types";
+import {
+  LOGIN_FAILURE,
+  LOGIN_REQUEST,
+  LOGIN_SUCCESS,
+  LOGOUT_REQUEST,
+  LOGOUT_FAILURE,
+  LOGOUT_SUCCESS,
+} from "./types";
 
 export const logIn = (user) => {
   console.log(user);
 
   return async (dispatch) => {
-    dispatch({type: LOGIN_REQUEST})
+    dispatch({ type: LOGIN_REQUEST });
     const res = await axios.post("/admin/signin", {
       ...user,
     });
@@ -13,12 +20,12 @@ export const logIn = (user) => {
     if (res.status === 200) {
       const { token, user } = res.data;
       localStorage.setItem("token", token);
-      localStorage.setItem('user', JSON.stringify(user))
+      localStorage.setItem("user", JSON.stringify(user));
       dispatch({
         type: LOGIN_SUCCESS,
         payload: {
           token,
-          user
+          user,
         },
       });
     } else {
@@ -33,30 +40,41 @@ export const logIn = (user) => {
 };
 
 export const isUserLoggedIn = () => {
-  return async dispatch => {
-    const token = localStorage.getItem('token');
-    if(token){
-      const user = JSON.parse(localStorage.getItem('user'));
+  return async (dispatch) => {
+    const token = localStorage.getItem("token");
+    if (token) {
+      const user = JSON.parse(localStorage.getItem("user"));
       dispatch({
         type: LOGIN_SUCCESS,
         payload: {
-          token, user
-        }
+          token,
+          user,
+        },
       });
-    }else{
+    } else {
       dispatch({
         type: LOGIN_FAILURE,
-        payload: {error: 'Failed to login'}
-      })
+        payload: { error: "Failed to login" },
+      });
     }
-  }
-}
+  };
+};
 
 export const signout = () => {
-  return async dispatch => {
-    localStorage.clear();
-    dispatch({
-      type: LOGOUT_REQUEST
-    })
-  }
-}
+  return async (dispatch) => {
+    dispatch({ type: LOGOUT_REQUEST });
+    const res = await axios.post("/admin/signout");
+
+    if (res.status === 200) {
+      localStorage.clear();
+      dispatch({
+        type: LOGOUT_SUCCESS,
+      });
+    } else {
+      dispatch({
+        type: LOGOUT_FAILURE,
+        payload: { error: res.data.error },
+      });
+    }
+  };
+};
